+++ lib/src/api/bridge/frb.dart
@@ -1,3 +1,4 @@
 export 'api.dart';
 export 'api/media_device_info.dart';
 export 'api/media_display_info.dart';
+export 'rtc_ice_candidate_stats.dart';
